# PERF: Try fast/slow paths only once in DataFrameGroupby.transform