--- a/doc/source/whatsnew/v1.4.0.rst
+++ b/doc/source/whatsnew/v1.4.0.rst
@@ -159,7 +159,7 @@ Deprecations
 Performance improvements
 ~~~~~~~~~~~~~~~~~~~~~~~~
 - Performance improvement in :meth:`.GroupBy.sample`, especially when ``weights`` argument provided (:issue:`34483`)
--
+- Performance improvement in :meth:`.GroupBy.transform` for user-defined functions (:issue:`41598`)
 
 .. ---------------------------------------------------------------------------