--- a/pandas/core/groupby/generic.py
+++ b/pandas/core/groupby/generic.py
@@ -1309,17 +1309,21 @@ class DataFrameGroupBy(GroupBy[DataFrame]):
         gen = self.grouper.get_iterator(obj, axis=self.axis)
         fast_path, slow_path = self._define_paths(func, *args, **kwargs)
 
+        path = None
         for name, group in gen:
             object.__setattr__(group, "name", name)
 
-            # Try slow path and fast path.
-            try:
-                path, res = self._choose_path(fast_path, slow_path, group)
-            except TypeError:
-                return self._transform_item_by_item(obj, fast_path)
-            except ValueError as err:
-                msg = "transform must return a scalar value for each group"
-                raise ValueError(msg) from err
+            if path is None:
+                # Try slow path and fast path.
+                try:
+                    path, res = self._choose_path(fast_path, slow_path, group)
+                except TypeError:
+                    return self._transform_item_by_item(obj, fast_path)
+                except ValueError as err:
+                    msg = "transform must return a scalar value for each group"
+                    raise ValueError(msg) from err
+            else:
+                res = path(group)
 
             if isinstance(res, Series):