@rhshadrach 2021-07-07
@jreback @jbrockmendel Updated and ready for review. 

I also tried _only_ using the slow path (instead of choosing between fast/slow); this resulted in a 0.7 ratio for the benchmark `groupby.TransformEngine.time_dataframe_cython(False)` as opposed to a 0.3 in this PR. While I'd prefer to find a way to not have to choose between different paths based on the first group, the performance seems worth it in my opinion.