@jbrockmendel 2021-07-07
> While I'd prefer to find a way to not have to choose between different paths based on the first group

IIUC you mean do some introspection on the function (or require users to specify somehow) to determine this instead of a try/except?  That would be nice.

> this resulted in a 0.7 ratio for the benchmark

What about for the groupby benchmarks more generally?  The slow path has been optimized a bunch this year (xref #40263)