@jreback 2021-07-07
I think its fine to determine how to call it based on the first group. we need *some* metric.