@rhshadrach 2021-07-07
@jbrockmendel 

> IIUC you mean do some introspection on the function (or require users to specify somehow) to determine this instead of a try/except? That would be nice.

Introspection seems fragile to me. I didn't consider user specification but I like the idea as long as it isn't required. Another option would be to restrict what UDFs are acceptable (e.g. must return Series, DataFrame, or NumPy array, and maybe some other things). For this latter option, we could then handle everything acceptable via the fastpath and remove the slowpath entirely. But perhaps the flexibility of current handling makes this an unwelcome trade-off.