@rhshadrach 2021-07-08
@jbrockmendel - This is slowpath only (L1385, first line of _choose_path is `return slow_path, slow_path(group)`):

```
       before           after         ratio
     [9ffc924e]       [462500f7]
     <whatsnew_contributing~2>       <groupby_transform_path>
+      17.9±0.6ms         25.4±1ms     1.42  groupby.Cumulative.time_frame_transform('int64', 'cummin')
+     21.3±0.07ms       26.5±0.5ms     1.24  groupby.Nth.time_frame_nth_any('datetime')
+      25.9±0.7ms       31.4±0.4ms     1.21  groupby.Nth.time_frame_nth_any('object')
+     11.9±0.09ms       14.0±0.3ms     1.18  groupby.Nth.time_frame_nth('object')
+      23.5±0.1ms         26.7±3ms     1.14  groupby.Nth.time_series_nth_any('datetime')
-      71.1±0.5ms       57.4±0.5ms     0.81  groupby.TransformEngine.time_dataframe_cython(True)
-      71.0±0.5ms       56.9±0.2ms     0.80  groupby.TransformEngine.time_dataframe_cython(False)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```